+++ src/main/java/wordle/model/TileService.java
@@ -1,4 +1,4 @@
-package wordle;
+package wordle.model;
 
 import java.util.List;
