+++ src/main/java/wordle/model/WordleValidator.java
@@ -1,4 +1,4 @@
-package wordle;
+package wordle.model;
 
 public class WordleValidator {
     private static final int VALID_SIZE = 5;
