+++ src/main/java/wordle/model/Words.java
@@ -1,4 +1,4 @@
-package wordle;
+package wordle.model;
 
 import java.time.LocalDate;
 import java.util.List;
