+++ src/test/java/wordle/LetterCounterTest.java
@@ -5,6 +5,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wordle.model.Letter;
+import wordle.model.LetterCounter;
+import wordle.model.Letters;
 
 class LetterCounterTest {
     @DisplayName("카운트를 감소시킬 수 있는 Letter만 필터링한다.")
