+++ src/test/java/wordle/LettersTest.java
@@ -1,13 +1,12 @@
 package wordle;
 
-import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wordle.model.Letter;
+import wordle.model.Letters;
 
 class LettersTest {
