+++ src/test/java/wordle/TileServiceTest.java
@@ -1,13 +1,13 @@
 package wordle;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
 import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import wordle.model.*;
 
 class TileServiceTest {
